# swagger/swagger.yaml
basePath: /
definitions:
...
  models.AuthError:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  models.ShortenURLRequest:
    properties:
      url:
        type: string
    required:
    - url
    type: object
  models.SupportRequest:
    properties:
      email:
        maxLength: 255
        type: string
      message:
        maxLength: 5000
        minLength: 1
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - email
    - message
    - name
    type: object
  models.SupportResponse:
    properties:
      error:
        $ref: '#/definitions/models.AuthError'
      message:
        type: string
      success:
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@sniply.co.in
    name: API Support
    url: https://sniply.co.in/support
  description: A URL shortener service API with authentication and analytics
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://sniply.co.in/terms
  title: URL Shortener API
  version: "1.0"
paths:
...
  /api/delete/{code}:
    delete:
      consumes:
      - application/json
      description: Deletes a shortened URL by its code (only if owned by the authenticated
        user)
      parameters:
      - description: Short code of the URL
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete a URL
      tags:
      - urls
  /api/shorten:
    post:
      consumes:
      - application/json
      description: Creates a shortened URL from a long URL
      parameters:
      - description: URL to shorten
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.ShortenURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Shorten a URL
      tags:
      - urls
  /api/support:
    post:
      consumes:
      - application/json
      description: Submits a support request with rate limiting
      parameters:
      - description: Support request details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.SupportRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SupportResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.SupportResponse'
        "429":
          description: Too Many Requests
          schema:
            additionalProperties: true
            type: object
      summary: Submit support request
      tags:
      - support
  /api/test:
    get:
      consumes:
      - application/json
      description: Returns the count of tables in the database
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Test database connection
      tags:
      - test
  /api/urls:
    get:
      consumes:
      - application/json
      description: Returns all shortened URLs created by the authenticated user with
        statistics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List user's URLs
      tags:
      - urls
  /api/urls/{code}/stats:
    get:
      consumes:
      - application/json
      description: Returns detailed statistics for a shortened URL (only if owned
        by the authenticated user)
      parameters:
      - description: Short code of the URL
        in: path
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get URL statistics
      tags:
      - urls
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"